Increase the size of tx packet counters from 32bit to 64bit

+++ XDMA/linux-kernel/xdma/alinx_arch.h
@@ -115,8 +115,8 @@ struct qav_state {
 };
 
 struct buffer_tracker {
-	uint32_t entry_count;
-	uint32_t last_tx_count;
+	uint64_t entry_count;
+	uint64_t last_tx_count;
 };
 
 struct tsn_config {
